// types/dewanPenguji.ts
export type DewanPenguji = {
    IdUsersDpkakp: number;
    NamaUsersDpkakp: string;
    TypeDpkapk: string;
    Nik: number;
    Alamat: string;
    Provinsi: string;
    Cities: string;
    Nip: string;
    AsalInstansi: string;
    NomorTelpon: string;
    Email: string;
    Jabatan: string;
    Golongan: string;
    Pendidikan: string;
    TipeKeahlian: string;
    Foto: string;
    PengalamanBerlayar: string;
    Ijazah: string;
    SertifikatKeahlian: string;
    SertifikatTot: string;
    SertifikatToe: string;
    SertifikatToeSimulator: string;
    SerifikatAuditor: string;
    SertifikatLainnya: string;
    BukuPelaut: string;
    Expreience: string | null;
  };
  